+++ WT_Handle.h
@@ -38,7 +38,7 @@ class WT_Handle
 		int Load(loadkit & s) { return fm->Load(s);};
 		int Save(savekit & s){ return fm->Save(s);};
 
-		int GetN(){ return fm->GetN();}
+		i64 GetN(){ return fm->GetN();}
 		int GetAlphabetsize(){return fm->GetAlphabetsize();}
 		int SizeInByte(){ return fm->SizeInByte();};
 		int SizeInByte_count() { return fm->SizeInByte_count();};
